Break up binding kubernetesobject tests to separate file for readability
